norminated and changed atoi to atol

+++ philo/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: gkhaishb <gkhaishb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/30 12:27:18 by gkhaishb          #+#    #+#             */
-/*   Updated: 2023/07/04 16:40:20 by gkhaishb         ###   ########.fr       */
+/*   Updated: 2023/07/06 10:36:43 by gkhaishb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ typedef struct s_philo
 	pthread_mutex_t	mutex_eating;
 }					t_philo;
 
-int			ft_atoi(const char *str);
+long long	ft_atoi(const char *str);
 void		ft_init_forks(t_data *data);
 int			ft_init_data(t_data *data, int argc, char **argv);
 void		ft_run(t_data *data, int num);
